chore(build): remove tslint options
We've removed tslint from build cycles a while ago, but these options remained.

Close #2616
