+++ packages/angular-cli/models/webpack-build-development.ts
@@ -1,17 +1,5 @@
 const path = require('path');
 
-import * as webpack from 'webpack';
-
-declare module 'webpack' {
-    export interface LoaderOptionsPlugin {}
-    export interface LoaderOptionsPluginStatic {
-        new (optionsObject: any): LoaderOptionsPlugin;
-    }
-    interface Webpack {
-        LoaderOptionsPlugin: LoaderOptionsPluginStatic;
-    }
-};
-
 export const getWebpackDevConfigPartial = function(projectRoot: string, appConfig: any) {
   return {
     devtool: 'cheap-module-source-map',
@@ -21,17 +9,6 @@ export const getWebpackDevConfigPartial = function(projectRoot: string, appConfi
       sourceMapFilename: '[name].map',
       chunkFilename: '[id].chunk.js'
     },
-    plugins: [
-      new webpack.LoaderOptionsPlugin({
-        options: {
-          tslint: {
-            emitErrors: false,
-            failOnHint: false,
-            resourcePath: path.resolve(projectRoot, appConfig.root)
-          },
-        }
-      })
-    ],
     node: {
       fs: 'empty',
       global: true,
